Extract client project
