+++ client/src/main/scala/client/parser/model/ParserStatus.scala
@@ -1,3 +1,3 @@
-package grasshopper.model
+package grasshopper.client.parser.model
 
 case class ParserStatus(status: String, upSince: String, time: String, host: String)
