+++ geocoder/src/main/scala/grasshopper/model/Status.scala
@@ -1,0 +1,3 @@
+package grasshopper.model
+
+case class Status(status: String, service: String, time: String, host: String)
