+++ geocoder/src/main/scala/grasshopper/protocol/GrasshopperJsonProtocol.scala
@@ -1,8 +1,8 @@
 package grasshopper.protocol
 
-import grasshopper.model.ParserStatus
+import grasshopper.model.Status
 import spray.json.DefaultJsonProtocol
 
 trait GrasshopperJsonProtocol extends DefaultJsonProtocol {
-  implicit val statusFormat = jsonFormat4(ParserStatus.apply)
+  implicit val statusFormat = jsonFormat4(Status.apply)
 }
